add shim for expand output partitioning
